(sandbox/malach_en) Adding online multisplice recipe (will still have to figure out the parameters)
git-svn-id: file:///home/jtrmal/Downloads/kaldi-git-conversion/kaldi-svn/sandbox/malach_en@4973 5e6a8d80-dfce-4ca6-a32a-6e07a63d50c8

+++ egs/malach_en/s5/run_uwb.sh
@@ -1,3 +1,4 @@
+#!/bin/bash
 #/export/corpora/LDC/LDC2012S05/data/
 #/export/corpora/LDC/LDC2012S05/data/
 train_audio_path=/export/corpora/LDC/LDC2012S05/data
@@ -5,6 +6,10 @@ train_trs_path=/export/a13/jtrmal/malach_uwb_sets/train/
 dev_audio_path=/export/corpora/LDC/LDC2012S05/data/mp2/
 dev_audio_path=/export/a13/jtrmal/malach_uwb_sets/test/
 dev_trs_path=/export/a13/jtrmal/malach_uwb_sets/test/
+
+
+. ./path.sh
+. ./cmd.sh
 
 set -e -o pipefail
 
@@ -14,46 +19,48 @@ if [[ `hostname -f` == *clsp.jhu.edu ]] ; then
   utils/create_split_dir.pl  /export/b0{1,2,3,4,5,6,7}/$w/${segpath}/storage \
     param/storage 2>/dev/null
 fi
-
-local/malach_prepare_data.sh \
-  --train-audio $train_audio_path --dev-audio $dev_audio_path \
-  --train-transcripts $train_trs_path --dev-transcripts $dev_trs_path
-
-
-local/malach_filter_transcripts_uwb.sh data/local/dev
-local/malach_filter_transcripts_uwb.sh data/local/train
-
-local/malach_create_lexicon.sh --nonspeech conf/nonspeech-uwb.txt \
-  --sil "<SILENCE>" --oov "<UNK>" \
-  data/local/train/transcriptions_for_lm.txt data/local/dict
-
-#local/malach_create_test_lexicon.sh --nonspeech conf/nonspeech-uwb.txt \
-#  --sil "<SILENCE>" --oov "<UNK>" \
-#  data/local/train/transcriptions_for_lm.txt data/local/dict
-
-local/malach_create_lang.sh --oov "<UNK>" \
-  data/local/dict data/local/lang data/lang
-
-#local/malach_create_lang.sh --oov "<UNK>" \
-#  data/local/dict_test data/local/lang_test data/lang_test
+if [ ! -f data/.done ] ; then
+  local/malach_prepare_data.sh \
+    --train-audio $train_audio_path --dev-audio $dev_audio_path \
+    --train-transcripts $train_trs_path --dev-transcripts $dev_trs_path
 
 
+  local/malach_filter_transcripts_uwb.sh data/local/dev
+  local/malach_filter_transcripts_uwb.sh data/local/train
 
-cp -r data/lang data/lang_test
-local/malach_create_lang_test.sh --oov "<UNK>" data/local/train/transcriptions_for_lm.txt \
-  data/local/srilm data/lang_test
+  local/malach_create_lexicon.sh --nonspeech conf/nonspeech-uwb.txt \
+    --sil "<SILENCE>" --oov "<UNK>" \
+    data/local/train/transcriptions_for_lm.txt data/local/dict
+
+  #local/malach_create_test_lexicon.sh --nonspeech conf/nonspeech-uwb.txt \
+  #  --sil "<SILENCE>" --oov "<UNK>" \
+  #  data/local/train/transcriptions_for_lm.txt data/local/dict
+
+  local/malach_create_lang.sh --oov "<UNK>" \
+    data/local/dict data/local/lang data/lang
+
+  #local/malach_create_lang.sh --oov "<UNK>" \
+  #  data/local/dict_test data/local/lang_test data/lang_test
 
 
-local/malach_create_kaldi_files.sh --map-oov false \
-  data/local/dev  data/lang  data/dev
-local/malach_create_kaldi_files.sh --map-oov true  \
-  data/local/train data/lang data/train
+
+  cp -r data/lang data/lang_test
+  local/malach_create_lang_test.sh --oov "<UNK>" data/local/train/transcriptions_for_lm.txt \
+    data/local/srilm data/lang_test
 
 
+  local/malach_create_kaldi_files.sh --map-oov false \
+    data/local/dev  data/lang  data/dev
+  local/malach_create_kaldi_files.sh --map-oov true  \
+    data/local/train data/lang data/train
+
+  touch data/.done
+fi
 
 for dataset in train dev ; do
   if [ ! -f data/$dataset/.done ] ; then
-    steps/make_mfcc.sh --nj 32 --cmd "$train_cmd" data/$dataset exp/make_param/$dataset param
+    #steps/make_plp_pitch.sh --nj 32 --cmd "$train_cmd" data/$dataset exp/make_param/$dataset param
+    steps/make_mfcc_pitch.sh --nj 32 --cmd "$train_cmd" data/$dataset exp/make_param/$dataset param
     utils/fix_data_dir.sh data/$dataset/
     steps/compute_cmvn_stats.sh data/$dataset exp/make_param/$dataset param/
     utils/fix_data_dir.sh data/$dataset
@@ -75,14 +82,19 @@ fi
 # Acoustic model parameters
 numLeavesTri1=1000
 numGaussTri1=10000
+
 numLeavesTri2=1000
 numGaussTri2=20000
+
 numLeavesTri3=6000
 numGaussTri3=75000
-numLeavesMLLT=6000
-numGaussMLLT=75000
-numLeavesSAT=6000
-numGaussSAT=75000
+
+numLeavesMLLT=8000
+numGaussMLLT=100000
+
+numLeavesSAT=10000
+numGaussSAT=150000
+
 numGaussUBM=800
 numLeavesSGMM=10000
 numGaussSGMM=80000
@@ -94,15 +106,15 @@ if [ ! -f exp/mono/.done ]; then
 fi
 
 (
-  [ -f exp/mono/decode_dev_uwb/.done ] && exit 0
+  [ -f exp/mono/decode_dev/.done ] && exit 0
 
   utils/mkgraph.sh --mono data/lang_test exp/mono/ exp/mono/graph
 
   steps/decode.sh  --cmd "$decode_cmd" \
     --parallel-opts "-pe smp 4" --num-threads 4 --nj 32 \
-    exp/mono/graph/ data/dev_uwb/ exp/mono/decode_dev_uwb
+    exp/mono/graph/ data/dev/ exp/mono/decode_dev
 
-  touch exp/mono/decode_dev_uwb/.done
+  touch exp/mono/decode_dev/.done
 ) &
 
 if [ ! -f exp/tri1/.done ]; then
@@ -120,15 +132,15 @@ fi
 
 
 (
-  [ -f exp/tri1/decode_dev_uwb/.done ] && exit 0
+  [ -f exp/tri1/decode_dev/.done ] && exit 0
 
   utils/mkgraph.sh data/lang_test exp/tri1/ exp/tri1/graph
 
   steps/decode_si.sh  --cmd "$decode_cmd" \
     --parallel-opts "-pe smp 4" --num-threads 4 --nj 32 \
-    exp/tri1/graph/ data/dev_uwb/ exp/tri1/decode_dev_uwb
+    exp/tri1/graph/ data/dev/ exp/tri1/decode_dev
 
-  touch exp/tri1/decode_dev_uwb/.done
+  touch exp/tri1/decode_dev/.done
 ) &
 
 echo ---------------------------------------------------------------------
@@ -145,15 +157,15 @@ if [ ! -f exp/tri2/.done ]; then
 fi
 
 (
-  [ -f exp/tri2/decode_dev_uwb/.done ] && exit 0
+  [ -f exp/tri2/decode_dev/.done ] && exit 0
 
   utils/mkgraph.sh data/lang_test exp/tri2/ exp/tri2/graph
 
   steps/decode_si.sh  --cmd "$decode_cmd" \
     --parallel-opts "-pe smp 4" --num-threads 4 --nj 32 \
-    exp/tri2/graph/ data/dev_uwb/ exp/tri2/decode_dev_uwb
+    exp/tri2/graph/ data/dev/ exp/tri2/decode_dev
 
-  touch exp/tri2/decode_dev_uwb/.done
+  touch exp/tri2/decode_dev/.done
 ) &
 
 
@@ -172,15 +184,15 @@ fi
 
 
 (
-  [ -f exp/tri3/decode_dev_uwb/.done ] && exit 0
+  [ -f exp/tri3/decode_dev/.done ] && exit 0
 
   utils/mkgraph.sh data/lang_test exp/tri3/ exp/tri3/graph
 
   steps/decode_si.sh  --cmd "$decode_cmd" \
     --parallel-opts "-pe smp 4" --num-threads 4 --nj 32 \
-    exp/tri3/graph/ data/dev_uwb/ exp/tri3/decode_dev_uwb
+    exp/tri3/graph/ data/dev/ exp/tri3/decode_dev
 
-  touch exp/tri3/decode_dev_uwb/.done
+  touch exp/tri3/decode_dev/.done
 ) &
 
 
@@ -200,15 +212,15 @@ if [ ! -f exp/tri4/.done ]; then
 fi
 
 (
-  [ -f exp/tri4/decode_dev_uwb/.done ] && exit 0
+  [ -f exp/tri4/decode_dev/.done ] && exit 0
 
   utils/mkgraph.sh data/lang_test exp/tri4/ exp/tri4/graph
 
   steps/decode.sh  --cmd "$decode_cmd" \
     --parallel-opts "-pe smp 4" --num-threads 4 --nj 32 \
-    exp/tri4/graph/ data/dev_uwb/ exp/tri4/decode_dev_uwb
+    exp/tri4/graph/ data/dev/ exp/tri4/decode_dev
 
-  touch exp/tri4/decode_dev_uwb/.done
+  touch exp/tri4/decode_dev/.done
 ) &
 
 echo ---------------------------------------------------------------------
@@ -226,15 +238,26 @@ if [ ! -f exp/tri5/.done ]; then
   touch exp/tri5/.done
 fi
 
+utils/mkgraph.sh data/lang_test  exp/tri5/ exp/tri5/graph
   (
-    [ -f exp/tri5/decode_dev_uwb_fmllr_extra/.done ] && exit 0
-
-    utils/mkgraph.sh data/lang_test  exp/tri5/ exp/tri5/graph
+    [ -f exp/tri5/decode_dev_fmllr_extra_A/.done ] && exit 0
+    
     steps/decode_fmllr_extra.sh  --cmd "$decode_cmd" \
       --parallel-opts "-pe smp 4" --num-threads 4 --nj 32 \
-      exp/tri5/graph/ data/dev_uwb/ exp/tri5/decode_dev_uwb_fmllr_extra
+      exp/tri5/graph/ data/dev_A/ exp/tri5/decode_dev_fmllr_extra_A
 
-    touch exp/tri5/decode_dev_uwb_fmllr_extra/.done
+
+    touch exp/tri5/decode_dev_fmllr_extra_A/.done
+  ) &
+  (
+  [ -f exp/tri5/decode_dev_fmllr_extra_B/.done ] && exit 0
+    
+    steps/decode_fmllr_extra.sh  --cmd "$decode_cmd" \
+      --parallel-opts "-pe smp 4" --num-threads 4 --nj 32 \
+      exp/tri5/graph/ data/dev_B/ exp/tri5/decode_dev_fmllr_extra_B
+
+
+    touch exp/tri5/decode_dev_fmllr_extra_B/.done
   ) &
 
 wait
